bot_client/requirements.txt
bcrypt==3.2.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
idna==3.3
pyaes==1.6.1
pycparser==2.21
pydantic==1.9.0
Pyrogram==2.0.16
PySocks==1.7.1
python-dotenv==0.20.0
requests==2.27.1
TgCrypto==1.2.3
typing_extensions==4.2.0
urllib3==1.26.9
